Started separation of API and implementation for KaitaiStream
